Read correct shadow frame sizes

+++ include/ztalib/data/ZtaFrame.h
@@ -27,7 +27,7 @@
 struct ZtaFrame
 {
     uint32_t frameSize; // in bytes
-    uint8_t height;
+    uint16_t height;
     uint8_t isShadow;
     uint16_t width;
     int16_t y; // y offset
